PET-124 Created controller superclass and refactored LoginController
