+++ src/main/java/org/controllers/Controller.java
@@ -1,0 +1,22 @@
+package org.controllers;
+
+import javafx.event.ActionEvent;
+import javafx.fxml.FXML;
+import javafx.fxml.FXMLLoader;
+import javafx.scene.Node;
+import javafx.scene.Parent;
+import javafx.scene.Scene;
+import javafx.stage.Stage;
+
+public class Controller {
+
+    @FXML
+    public void redirectToRegister(ActionEvent event) throws Exception{
+        Node node = (Node) event.getSource();
+        Stage currentStage = (Stage) node.getScene().getWindow();
+        Parent root = FXMLLoader.load(getClass().getClassLoader().getResource("registerScene.fxml"));
+        currentStage.setTitle("Register");
+        currentStage.setScene(new Scene(root, 800, 600));
+        currentStage.show();
+    }
+}
